// src/app/models/mae-centro-costos.ts
export class MaeCentroCostrosDTO {
    codEmpresa?: string;
    numVerCCostos?: string;
    codCCostos?: string;
    codCCostosSup?: string;
    desCCostos?: string;
    desAbreviatura?: string;
    numNivel?: number;
    codCuentaPresup?: string;
    indBaja?: string;
    fecBaja?: Date;
    codUserActual?: string;
    fecActualiza?: Date;
    indEquipos?: string;
    codEquivalencia?: string;
    codSucursal?: string;
    codAlmacen?: string;
    tipCCostos?: string;
    flgNivelOrg?: string;
    codZona?: string;
    codEquivalencia2?: string;
    tipGrupoEmp?: string;
    tipGrupoObr?: string;
    tipGestionOb?: string;
    codCCostosGroup?: string;
    desMarca?: string;
    desModelo?: string;
    desSerie?: string;
    codAnio?: string;
    indAuxiliar?: string;
    desMotor?: string;
    desPlaca?: string;
    indAc?: string;
    responsable?: string;
    direccion?: string;
    horario?: string;
    ubicacion?: string;
    distrito?: string;
    codCia?: string;
    indAlmStk?: string;
    fecCreacion?: Date;
    codUserCrea?: string;
    hijos?: MaeCentroCostrosDTO[];
}