// src/app/models/mae-banco.ts
export class MaeBanco {
    codEmpresa?: string;
    codAuxiliar?: string;
    desAuxiliar?: string;
    numRuc?: string;
    numTelefono?: string;
    desDireccion?: string;
    codPais?: string;
    tipPersoneria?: string;
    codUbigeo?: string;
    indReten6?: string;
    indBuenContrib?: string;
}