// src/app/models/mae-documento.ts
export class MaeDocumento {
    codDocumento?: string;
    desDocumento?: string;
    indBaja?: string;
    fecBaja?: Date;
    desCorta?: string;
    codSunat?: string;
    indSistema?: string;
    codEmpresa?: string;
    numVerPlanCuentas?: string;
    codCuentaSoles?: string;
    codCuentaDolares?: string;
    indNc?: string;
    indDebeHaber?: string;
    indReten6?: string;
    indExon?: string;
    indDaot?: string;
    indExp?: string;
    tipCuentaRel?: string;
    codCuentaSolesRel?: string;
    codCuentaDolaresRel?: string;
    ind4ta?: string;
    indNb?: string;
    indIsc?: string;
    indDua?: string;
    longitudSerie?: number;
    longitudNumero?: number;
}